Aggiunto file xml per i 45 punti di interesse. Aggiornata libreria ADT.Graphs. Aggiornate classi Configuraion, Solver, SolverValidator, SolverEvaluator, Algorithm, RunningConfigs. Aggiunta lista degli algorithmTasks da attendere negli algoritmi di ricerca locale. Bug fixing nella classe NearestNeighborAlgorithm per la selezione di un nodo casuale. Aggiornata classe LocalSearchAlgorithm.

+++ CityScover.Engine/Configuration.cs
@@ -3,7 +3,7 @@
 // Version 1.0
 //
 // Authors: Andrea Ritondale, Andrea Mingardo
-// File update: 24/09/2018
+// File update: 29/09/2018
 //
 
 using CityScover.Entities;
@@ -30,7 +30,7 @@ namespace CityScover.Engine
       internal TourCategoryType TourCategory { get; set; }
       internal int StartPOIId { get; set; }
       internal int PointsCount { get; set; }
-      internal float WalkingSpeed { get; set; }
+      internal double WalkingSpeed { get; set; }
       internal DateTime ArrivalTime { get; set; }
       internal TimeSpan TourDuration { get; set; }
       internal ProblemType CurrentProblem { get; set; }
